Implement q_free function
The previous q_free function was empty and unable to free the space
used by the queue.

This new implementation traverses all list elements, uses the
container_of macro to find the structure containing the list_head node,
and frees the space for each element_t and its inner value. Finally, it
frees the head node.

Change-Id: I3f398187d482658738d546ada8be433fb13ea630