Implement q_remove_head, q_remove_tail functions
The previous implementation did not provide functions for removing
 elements from the head or tail of the queue, simply returns a null
 pointer.

The q_remove_head function removes the element from the head of the
 queue and, if a valid buffer is provided, safely copies the element’s
 value into the buffer using strncpy, ensuring no buffer overflow
occurs. Similarly, the q_remove_tail function removes the element from
the tail of the queue and copies value to buffer safely.

The new implementation ensure that the element is properly removed from
the list and that memory errors are avoided by using strncpy to limit
the number of characters copied into the provided buffer, and ensuring
the string is null-terminated.

Change-Id: I24432ce42c73b4aefcbe54eda2b4568078710b23